Adding a 2nd property on the replace benchmark
