--- java/java-ranger-regression/replace5_eqchk.yml
@@ -1,10 +1,0 @@
-format_version: "2.0"
-input_files:
-  - ../common/
-  - replace5_eqchk/
-properties:
-  - property_file: ../properties/assert.prp
-    expected_verdict: false
-
-options:
-  language: Java
